back_end/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
python-dotenv
flask-uploads